+++ lib/oli/rendering/content.ex
@@ -7,6 +7,7 @@ defmodule Oli.Rendering.Content do
 
   alias Oli.Rendering.Context
   alias Oli.Utils
+  alias Oli.Rendering.Utils, as: RenderUtils
 
   require Logger
 
@@ -41,7 +42,8 @@ defmodule Oli.Rendering.Content do
   @callback code_line(%Context{}, next, %{}) :: [any()]
   @callback blockquote(%Context{}, next, %{}) :: [any()]
   @callback a(%Context{}, next, %{}) :: [any()]
-  @callback popup(%Context{}, next, %{}) :: [any()]
+  @callback popup_anchor(%Context{}, next, %{}, String.t()) :: [any()]
+  @callback popup_content(%Context{}, next, %{}, String.t()) :: [any()]
   @callback error(%Context{}, %{}, {Atom.t(), String.t(), String.t()}) :: [any()]
 
   @doc """
@@ -173,7 +175,20 @@ defmodule Oli.Rendering.Content do
         writer.a(context, next, element)
 
       "popup" ->
-        writer.popup(context, next, element)
+        id = UUID.uuid4()
+
+        [
+          writer.popup_anchor(context, next, element, id),
+          writer.popup_content(
+            context,
+            fn -> render(context, element["content"], writer) end,
+            element["content"],
+            id
+          )
+        ]
+
+      # content = RenderUtils.parse_html_content(element["content"], context)
+      # writer.popup(context, next, element, content)
 
       _ ->
         error_id = Utils.generate_error_id()